+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/speed/model/SpeedUnit.kt
@@ -1,0 +1,21 @@
+package com.mapbox.navigation.base.speed.model
+
+/**
+ * Specifies posted and current speed units.
+ */
+enum class SpeedUnit {
+    /**
+     * Posted speed in miles per hour
+     */
+    MILES_PER_HOUR,
+
+    /**
+     * Posted speed in kilometers per hour
+     */
+    KILOMETERS_PER_HOUR,
+
+    /**
+     * Current speed in meters per second
+     */
+    METERS_PER_SECOND
+}