+++ libnavui-speedlimit/src/main/java/com/mapbox/navigation/ui/speedlimit/SpeedLimitProcessor.kt
@@ -1,9 +1,20 @@
 package com.mapbox.navigation.ui.speedlimit
 
+import com.mapbox.navigation.base.formatter.UnitType
+import com.mapbox.navigation.base.speed.model.SpeedUnit
+import com.mapbox.navigation.ui.speedlimit.model.SpeedData
+
 internal class SpeedLimitProcessor {
 
     fun process(action: SpeedLimitAction): SpeedLimitResult {
-        return calculateSpeedLimitUpdate(action as SpeedLimitAction.CalculateSpeedLimitUpdate)
+        return when (action) {
+            is SpeedLimitAction.CalculateSpeedLimitUpdate -> {
+                calculateSpeedLimitUpdate(action)
+            }
+            is SpeedLimitAction.FindPostedAndCurrentSpeed -> {
+                getPostedAndCurrentSpeed(action)
+            }
+        }
     }
 
     private fun calculateSpeedLimitUpdate(
@@ -13,6 +24,44 @@ internal class SpeedLimitProcessor {
             action.speedLimit.speedKmph ?: 0,
             action.speedLimit.speedLimitUnit,
             action.speedLimit.speedLimitSign
+        )
+    }
+
+    private fun getPostedAndCurrentSpeed(
+        action: SpeedLimitAction.FindPostedAndCurrentSpeed
+    ): SpeedLimitResult {
+        val locationMatcher = action.locationMatcherResult
+        val currentSpeedMetersPerSecond: Double = locationMatcher.enhancedLocation.speed.toDouble()
+        val formattedCurrentSpeed = action.formatter.format(
+            SpeedData(
+                currentSpeedMetersPerSecond,
+                SpeedUnit.METERS_PER_SECOND,
+                action.distanceFormatterOptions.unitType
+            )
+        )
+
+        val formattedPostedSpeed = locationMatcher.speedLimit?.speedKmph?.let { speedKmph ->
+            action.formatter.format(
+                SpeedData(
+                    speedKmph.toDouble(),
+                    SpeedUnit.KILOMETERS_PER_HOUR,
+                    action.distanceFormatterOptions.unitType
+                )
+            )
+        }
+
+        val postedSpeedUnit = when (action.distanceFormatterOptions.unitType) {
+            UnitType.METRIC -> SpeedUnit.KILOMETERS_PER_HOUR
+            UnitType.IMPERIAL -> SpeedUnit.MILES_PER_HOUR
+        }
+
+        val speedSign = locationMatcher.speedLimit?.speedLimitSign
+
+        return SpeedLimitResult.PostedAndCurrentSpeed(
+            formattedPostedSpeed,
+            formattedCurrentSpeed,
+            postedSpeedUnit,
+            speedSign,
         )
     }
 }